moved send email to a different component
