+++ Server/package.json
@@ -18,6 +18,8 @@
     "express": "^4.21.1",
     "express-session": "^1.18.1",
     "mongoose": "^8.7.0",
+    "multer": "^1.4.5-lts.1",
+    "nodemailer": "^6.9.16",
     "pdf-lib": "^1.17.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
@@ -31,6 +33,4 @@
     "mongodb-memory-server": "^10.1.2",
     "supertest": "^7.0.0"
   }
-
 }
-
